moved ParsedFlowFyFile class to parsed_fy_file.py
