+++ library/src/domain/parsed_fy_file.py
@@ -1,7 +1,10 @@
 from enum import Enum
 from pathlib import Path
+from typing import Any
 
-from pydantic import BaseModel
+from pydantic import BaseModel, model_validator
+
+from domain.flow_template_model import FlowTemplateModel
 
 
 class ParsedFyFileKind(Enum):
@@ -12,3 +15,14 @@ class ParsedFyFile(BaseModel):
     file_type: ParsedFyFileKind
     output_py_file_path: Path
     template_model: BaseModel
+
+
+class ParsedFlowFyFile(ParsedFyFile):
+    template_model: FlowTemplateModel
+
+    @model_validator(mode="before")
+    @classmethod
+    def set_kind(cls, data: Any) -> Any:
+        if isinstance(data, dict):
+            data["kind"] = ParsedFyFileKind.FLOW
+        return data
